% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimamodel.R
\name{ARIMAModel}
\alias{ARIMAModel}
\title{Class ARIMAModel}
\description{
Represents ARIMA models.
}
\section{Direct Superclass}{
{
\itemize{
\item \link{FitModel}
}
}
}

\section{Important Dependencies}{
 {
\itemize{
\item \href{http://pkg.robjhyndman.com/forecast/reference/auto.arima.html}{forecast::auto.arima}
\item \href{http://pkg.robjhyndman.com/forecast/reference/forecast.html}{forecast::forecast}
}
}
}

\section{Public methods}{
{
\itemize{
  \item \strong{buildModel(...)} passes on parameters \code{stepwise} and \code{approximation} and fitting and forecasting functions to \link{FitModel}'s method \code{setFittedFcasted}; applies Box-Cox transformation and bias-adjustment; passes function which checks whether residuals are not independently distributed; allows generating forecast intervals from bootstrapped residuals if necessary
  \item \strong{useModel(fcast_period, ...)} calls \link{FitModel}'s method \code{executeUseModel} to fit model and forecast \code{fcast_period}-step ahead future; passes on the same fitting and forecasting functions as \code{buildmodel()}
}
}
}

\keyword{internal}
