% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvssforecaster.R
\name{CVSSForecaster}
\alias{CVSSForecaster}
\title{CVSS Forecaster}
\description{
Forecasts monthly mean CVSS scores of subsets of CWE vulnerabilities.
}
\section{Usage}{

\preformatted{
## Create new forecasting object.
cf <- CVSSForecaster$new()
}
}

\section{Public methods}{
{
\itemize{
  \item \strong{$new()} creates new CVSSForecaster object
  \item \strong{$getCWENames()} ...
  \item \strong{$getSeries(} ...
  \item \strong{$getBenchmark()} ...
  \item \strong{$getARIMA()} ...
  \item \strong{$getETS()} ...
  \item \strong{$getTSLinear()} ...
  \item \strong{$getNNAR()} ...
  \item \strong{$getARFIMA()} ...
  \item \strong{$getBaggedETS()} ...
  \item \strong{$getTBATS()} ...
  \item \strong{$getStructTS()} ...
  \item \strong{$getPlots()} ...
  \item \strong{$getAssessments()} ...
  \item \strong{$getAllAssessments()} ...
  \item \strong{$getBestModelList()} ...
  \item \strong{$getBestAssessments()} ...
  \item \strong{$useBest()} ...
  \item \strong{$plotUseBest()} ...
  \item \strong{$assessUseBest()} ...
  \item \strong{$setBenchmark()} ...
  \item \strong{$setARIMA()} ...
  \item \strong{$setETS()} ...
  \item \strong{$setTSLinear()} ...
  \item \strong{$setNNAR()} ...
  \item \strong{$setARFIMA()} ...
  \item \strong{$setBaggedETS()} ...
  \item \strong{$setTBATS()} ...
  \item \strong{$setStructTS()} ...
  \item \strong{$setCWEs()} ...
  \item \strong{$setBestModelList()} ...
}
}
}

