% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{CWE}
\alias{CWE}
\title{CWE data handler}
\description{
Manage the data that forms the basis of the forecasts
}
\section{Usage}{

\preformatted{
## Create new CWE object.
cf <- CWE$new()
}
}

\section{Public methods}{
{
\itemize{
  \item \strong{$new()} creates new CWE object
  \item \strong{$getStartYear()} item
  \item \strong{$getEndYear(plus_one = F)}
  \item \strong{$getStartMonth()} ...
  \item \strong{$getEndMonth()} ...
  \item \strong{$getBaseData()} ...
  \item \strong{$getTimeSeriesData()} ...
  \item \strong{$setStartYear()} ...
  \item \strong{$setEndYear()} ...
  \item \strong{$setStartMonth()} ...
  \item \strong{$setEndMonth()} ...
  \item \strong{$setBaseData()} ...
  \item \strong{$setTimeSeriesData()} ...
  \item \strong{$getAnalysisData()} ...
  \item \strong{$getMostInstances()} ...
  \item \strong{$getMostCritical()} ...
  \item \strong{$getChanging()} ...
  \item \strong{$getInterestingMonthlyData()} ...
  \item \strong{$getMonthlyData()} ...
}
}
}

