% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser.R
\name{cwes_highest_instances_change}
\alias{cwes_highest_instances_change}
\title{CWE categories with highest change-compared-to-the-last-year finder}
\usage{
cwes_highest_instances_change(data, period_threshold = 200)
}
\arguments{
\item{data}{A data.table with "published", "cwe" and "cvss_score" columns}

\item{period_threshold}{A number. Each CWE vulnerability included in the
output of this function must have the number of found vulnerabilities
(summed over the years of the time period) bigger than
the \code{period_threshold}.}
}
\value{
A data.table with columns "cwe" and "avg_change"
}
\description{
Finds yearly mean of absolute CVSS score percent change with function
\code{find_yearly_mean_change} and outputs the results that are above
the median of the change number or equal to the median.
}
\examples{
cwes_highest_instances_change(get_cwe_analysis_vulns())
}
