% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser.R
\name{cwes_most_critical_impact}
\alias{cwes_most_critical_impact}
\title{CWE categories with most critical impact finder}
\usage{
cwes_most_critical_impact(data, min_score = 4)
}
\arguments{
\item{data}{A data.table with "published", "cwe" and "cvss_score" columns}

\item{min_score}{A CVSS score number set to be the minimum that is of the
interest}
}
\value{
A data.frame with columns "cwe" and
"cvss" (summed year's average CVSS score that is divided by the length of
the given time period).
}
\description{
Finds CWE categories, which have a higher or equal yearly
summed average CVSS score than the parameter \code{min_score}.
}
\examples{
cwes_most_critical_impact(get_cwe_analysis_vulns())
}
