% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser.R
\name{cwes_most_instances}
\alias{cwes_most_instances}
\title{CWE categories with most instances finder}
\usage{
cwes_most_instances(data, threshold = 100)
}
\arguments{
\item{data}{A data.table with "published", "cwe" and "cvss_score" columns}

\item{threshold}{A number. Each CWE vulnerability included in the output
of this function must have the number of found vulnerabilities
(in any of the years of the time period) bigger than the \code{threshold}.}
}
\value{
A data.frame with columns "cwe" and "sum" (sum of the count of
vulnerabilities of the given time period)
}
\description{
Finds CWE categories, which have been discovered more times than the
threshold parameter's value in any of the years of the given time period.
}
\examples{
cwes_most_instances(get_cwe_analysis_vulns())
}
