% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{date_formatter}
\alias{date_formatter}
\title{Year or Year and Month formatter}
\usage{
date_formatter(dates, monthly)
}
\arguments{
\item{dates}{Calendar Date of class "Date"}

\item{monthly}{Logical parameter indicating whether month and year format
is required.}
}
\value{
If parameter \code{monthly} is TRUE, then
\code{zoo::yearmon}'s months and years. Otherwise the number of the year
in the dates.
}
\description{
Helps function \code{get_timegrouped_vulns} to group vulnerabilities
by months or by years by providing formatting.
}
\examples{
\dontrun{
date_formatter(c(as.Date("2011-01-01"),as.Date("2012-01-01")),monthly=TRUE)
}
}
\keyword{internal}
