% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{discern_types}
\alias{discern_types}
\title{One-CWE and multi-CWE separator}
\usage{
discern_types(data)
}
\arguments{
\item{data}{A data.table with columns "published", "cwe" and "cvss_score".}
}
\value{
A list with two data.table elements: one containing entries
with simple CWEs, other entries with multiple CWEs.
}
\description{
Separates one-CWE and multi-CWE vulnerability entries.
}
\examples{
\dontrun{discern_types(get_type_score_between(2011, 2012))}
}
\keyword{internal}
