% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{fill_missing_months_with_zero}
\alias{fill_missing_months_with_zero}
\title{Missing months filler}
\usage{
fill_missing_months_with_zero(start_end, vulns_by_month, selected_vulns_cwes)
}
\arguments{
\item{start_end}{A list containing the minimum and maximum year of the time
period}

\item{vulns_by_month}{A data.table output from
function \code{get_timegrouped_vulns(data, monthly = TRUE)}}

\item{selected_vulns_cwes}{A character vector with CWEs}
}
\value{
A data.table where all CWEs have a count for each month of the
selected time period.
}
\description{
Generates the missing months with zeros
}
\examples{
\dontrun{
vulns_by_month <- get_timegrouped_vulns(data, monthly = TRUE)
vulns_by_month <-
  fill_missing_months_with_zero(list(start_year, end_year), vulns_by_month,
                                  interesing_cwes)
}
}
\keyword{internal}
