% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{fill_missing_years_with_zero}
\alias{fill_missing_years_with_zero}
\title{Missing years filler}
\usage{
fill_missing_years_with_zero(start_end, vulns_by_year, selected_vulns_cwes)
}
\arguments{
\item{start_end}{A list containing \code{min} and \code{max} year from
\code{vulns_by_year$published}.}

\item{vulns_by_year}{A data.table output from
function \code{get_timegrouped_vulns}}

\item{selected_vulns_cwes}{A character vector with CWEs}
}
\value{
A data.table where all CWEs have a count for each year of the
selected time period.
}
\description{
Takes in the data of yearly vulnerability entries, for each selected CWE,
finds missing years and adds vulnerability count of 0 to there.
}
\examples{
\dontrun{
fill_missing_years_with_zero(list(overall_first_year, overall_last_year),
vulns_by_year,
cwe_year_above_200)
}
}
\keyword{internal}
