% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{find_yearly_mean_change}
\alias{find_yearly_mean_change}
\title{Percent Change of instances (last year vs all others) finder}
\usage{
find_yearly_mean_change(selected_vulns_cwes, vulns_by_year, last_year)
}
\arguments{
\item{selected_vulns_cwes}{A character vector with CWEs}

\item{vulns_by_year}{Output from function \code{fill_missing_years_with_zero}
or output \code{get_timegrouped_vulns}}

\item{last_year}{The given time period's final year.}
}
\value{
A data.table with columns "cwe" and "avg_change".
}
\description{
|(Percent Change)| = |((New Amount) - (Base Amount)) / (Base Amount)|,
where (New Amount) is last year vulnerability instances count and
(Base Amount) is each of other years' vulnerability instances counts.
Finally, calculates a mean of the found percent change numbers for a
a given CWE.
}
\examples{
\dontrun{
find_yearly_mean_change(cwes_above_threshold, vulns_by_year_all, last_year)
}
}
\keyword{internal}
