% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{findBestColumn}
\alias{findBestColumn}
\title{Find best column}
\usage{
findBestColumn(metrics)
}
\arguments{
\item{metrics}{forecast accuracy metrics}
}
\value{
the name of the selected column
}
\description{
Find best column
}
