% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\name{FitModel}
\alias{FitModel}
\title{Class FitModel}
\description{
Represents models that need to be fitted before forecasting.
}
\section{Direct Superclass}{
{
\itemize{
\item \link{NVDModel}
}
}
}

\section{Direct Subclasses}{
{
\itemize{
\item \link{ARFIMAModel},
\item \link{ARIMAModel},
\item \link{BaggedETSModel},
\item \link{ETSModel},
\item \link{NNARModel},
\item \link{StructTSModel},
\item \link{TBATSModel},
\item \link{TSLinearModel},
}
}
}

\section{Private attributes}{
{
\itemize{
  \item \strong{fit} the attribute to store the fitted model's object (default NA)
}
}
}

\section{Public methods}{
{
\itemize{
  \item \strong{getFitted()} returns private attribute \code{fit}
  \item \strong{setFitted(fit_result)} sets \code{fit_result} as the value of private attribute \code{fit}
  \item \strong{setFittedFcasted(fit_function, test_randomness, fcast_function, ...)} calls the fitting function, the forecasting function, performs residuals analysis and generates forecast intervals from bootstrapped residuals for ARIMA and ETS models when \code{!super$isPiIgnored() & super$areResidualsNotNormal() & !super$areResidualsNotRandom()}
  \item \strong{executeUseModel(fit_function, fcast_function, fcast_period, ...)} calls the fitting function, which uses the new training set (obtained by calling \link{NVDModel}'s method \code{getMergedTrainTestSet}), then calls the forecasting function which uses generating forecasting intervals from bootstrapped residuals when \code{!super$isBootstrapNotUsed() = TRUE}
}
}
}

\keyword{internal}
