% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser.R
\name{get_cwe_analysis_vulns}
\alias{get_cwe_analysis_vulns}
\title{CWE analysis data organiser}
\usage{
get_cwe_analysis_vulns(start_year = 2011, end_year = 2016)
}
\arguments{
\item{start_year}{A number inidicating the starting year of the
time period (min 2011)}

\item{end_year}{A number indicating the ending year of the
time period (max 2016)}
}
\value{
A data.table with "published", "cwe" and "cvss_score" columns
}
\description{
Gets the vulnerability data of the given time period. Makes sure all the
CWEs are separated.
}
\examples{
get_cwe_analysis_vulns(2012,2014)
}
