% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser.R
\name{get_interesting_cwes}
\alias{get_interesting_cwes}
\title{Interesting CWE categories finder}
\usage{
get_interesting_cwes(start_year = 2011, end_year = 2016, threshold = 100,
  min_score = 4, period_threshold = 200, as_monthly_ts = T)
}
\arguments{
\item{start_year}{A number inidicating the starting year of the
time period (min 2011 and default 2011).}

\item{end_year}{A number indicating the ending year of the
time period (max 2016 and default 2016).}

\item{threshold}{A number. Each CWE vulnerability included in the output
of \code{cwes_most_instances} function must have the number of
found vulnerabilities (default 100).
(in any of the years of the time period) bigger than the \code{threshold}.}

\item{min_score}{A CVSS score number set to be the minimum that is of the
interest (default 4.0).}

\item{period_threshold}{A number. Each CWE vulnerability included in the
output of function \code{cwes_highest_instances_change} must have the
number of found vulnerabilities.
(summed over the years of the time period) bigger than
the \code{period_threshold} (default 200).}

\item{as_monthly_ts}{Logical value. When TRUE}
}
\value{
A character vector of CWE names (default). If
parameter \code{as_monthly_ts} is TRUE, then a monthly time-series
object <"mts" "ts" "matrix"> of average CVSS scores of vulnerabilities with
CWEs in columns are returned instead.
}
\description{
Merges the CWEs found with functions \code{cwe_most_instances},
\code{cwes_most_critical_impact} and
\code{cwes_highest_instances_change} and returns the list of the found CWEs.
}
\examples{
get_interesting_cwes()
get_interesting_cwes(start_year=2014, end_year=2015, as_monthly_ts = TRUE)
}
