% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser.R
\name{get_nvd_entries}
\alias{get_nvd_entries}
\title{NVD entry extractor from XML file(s)}
\usage{
get_nvd_entries(files)
}
\arguments{
\item{files}{One or many XML Version 2.0 files from
\url{https://nvd.nist.gov/vuln/data-feeds#CVE_FEED}}
}
\value{
A data.table containing NVD entries extracted from (a) file(s).
}
\description{
Takes one or many files as arguments, processes these and returns a
data.table containing entries, each having an CVE id,
CWE types, logical value indicating whether the entry has been rejected,
publication date, CVSS base score, CVSS vectors (access vector,
access complexity, authentication, confidentiality impact, integrity impact,
availability impact)
}
\examples{
\dontrun{
get_nvd_entries(c("nvdcve-2.0-2013.xml","nvdcve-2.0-2015.xml"))
get_nvd_from_file("nvdcve-2.0-2013.xml")
}
}
