% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser.R
\name{get_nvd_from_file}
\alias{get_nvd_from_file}
\title{NVD entry extractor from one XML file}
\usage{
get_nvd_from_file(file)
}
\arguments{
\item{file}{An XML Version 2.0 file from
\url{https://nvd.nist.gov/vuln/data-feeds#CVE_FEED}}
}
\value{
A data.table containing NVD entries extracted from a file.
}
\description{
A helper function for \code{get_nvd_entries}, uses \code{XML::xmlEventParse}.
}
\examples{
\dontrun{get_nvd_from_file("nvdcve-2.0-2013.xml")}
}
\keyword{internal}
