% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser.R
\name{get_nvd_from_files}
\alias{get_nvd_from_files}
\title{NVD entry extractor from many XML files}
\usage{
get_nvd_from_files(files)
}
\arguments{
\item{files}{XML Version 2.0 files from
\url{https://nvd.nist.gov/vuln/data-feeds#CVE_FEED}}
}
\value{
A data.table containing NVD entries extracted from files.
}
\description{
A helper function for \code{get_nvd_entries}.
}
\examples{
\dontrun{get_nvd_entries(c("nvdcve-2.0-2013.xml","nvdcve-2.0-2015.xml"))}
}
\keyword{internal}
