% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{get_timegrouped_vulns}
\alias{get_timegrouped_vulns}
\title{CWE yearly or monthly characteristics creator}
\usage{
get_timegrouped_vulns(data, monthly = FALSE)
}
\arguments{
\item{data}{A data.table with columns "published", "cwe" and "cvss_score".}

\item{monthly}{Logical parameter. If TRUE, then monthly grouping is done,
otherwise the data is grouped yearly.}
}
\value{
A data.frame with columns "published", "cwe", "cvss_sum", "count"
and "avg_cvss".
}
\description{
For each CWE, finds the year's or month's sum of CVSS score,
the year's or month's count of the
CWE and calculates the year's or month's average CVSS score.
}
\examples{
get_timegrouped_vulns(get_cwe_analysis_vulns(2011, 2012))
}
\keyword{internal}
