% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{get_type_score_between}
\alias{get_type_score_between}
\title{Binary NVD data fetcher}
\usage{
get_type_score_between(start_year, end_year)
}
\arguments{
\item{start_year}{A number inidicating the starting year of the
time period (min 2011)}

\item{end_year}{A number indicating the ending year of the
time period (max 2016)}
}
\value{
A data.table with columns "published", "cwe", "cvss_score"
}
\description{
Takes a subset of columns from binary NVD vulnerability data stored inside
the package. Makes sure the vulnerabilities are published between
given time limits.
}
\examples{
get_type_score_between(2011,2016)
}
\keyword{internal}
