% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser_utils.R
\name{get_xml_base_metrics}
\alias{get_xml_base_metrics}
\title{CVSS base metrics node extractor from XML}
\usage{
get_xml_base_metrics(node)
}
\arguments{
\item{node}{An XMLNode \code{entry}.}
}
\value{
XMLNode \code{cvss:base_metrics}.
}
\description{
Gets XMLNode \code{cvss:base_metrics} which is
inside XMLNode \code{vuln:cvss}.
}
\examples{
\dontrun{get_xml_base_metrics(node)}
}
\keyword{internal}
