% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser_utils.R
\name{get_xml_cvss}
\alias{get_xml_cvss}
\title{CVSS metrics extractor from XML}
\usage{
get_xml_cvss(node)
}
\arguments{
\item{node}{An XMLNode \code{entry}.}
}
\value{
A list with CVSS metrics as characters.
}
\description{
Gets CVSS metrics from XML.
}
\examples{
\dontrun{get_xml_cvss(node)}
}
\keyword{internal}
