% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser_utils.R
\name{get_xml_cwes}
\alias{get_xml_cwes}
\title{CWE extractor from XML}
\usage{
get_xml_cwes(node)
}
\arguments{
\item{node}{An XMLNode \code{entry}.}
}
\value{
The character of CWEs from XML.
}
\description{
Extracts CWE types from XML node. If some NVD entry has
multiple types, then these are handled as one string connceted by "|". If
the types are missing, then NA is returned.
}
\examples{
\dontrun{get_xml_cwes(node)}

}
\keyword{internal}
