% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser_utils.R
\name{get_xml_date}
\alias{get_xml_date}
\title{Published date extractor from XML}
\usage{
get_xml_date(node)
}
\arguments{
\item{node}{An XMLNode \code{entry}.}
}
\value{
A character of date containing year, month and day.
}
\description{
Gets the datetime that shows when NVD entry was published and extracts a
date.
}
\examples{
\dontrun{get_xml_date(node)}
}
\keyword{internal}
