% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser_utils.R
\name{get_xml_metrics}
\alias{get_xml_metrics}
\title{Base metric's submetric extractor from XML}
\usage{
get_xml_metrics(base_metrics, metric)
}
\arguments{
\item{base_metrics}{Outcome from \code{get_xml_base_metrics}.}

\item{metric}{A string specifing the metric.}
}
\value{
A base metric value as a character.
}
\description{
Gets specific submetricts from \code{cvss:base_metrics}.
}
\examples{
\dontrun{get_xml_metrics(base_metrics, "score")}
}
\keyword{internal}
