% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser_utils.R
\name{get_xml_reject}
\alias{get_xml_reject}
\title{Rejected entry detector from XML}
\usage{
get_xml_reject(node)
}
\arguments{
\item{node}{An XMLNode \code{entry}.}
}
\value{
Logical value, TRUE when the entry is rejected.
}
\description{
Checks the XMLNode \code{entry}'s child \code{vuln:summary}. If the rejection
character is found, then TRUE is returned.
}
\examples{
\dontrun{get_xml_reject(node)}
}
\keyword{internal}
