% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nvd}
\alias{nvd}
\title{Reported Vulnerabilities's Data from National Vulnerability Database (NVD)}
\format{A data.table::data.table with 11 variables and 41190 rows:
\describe{
  \item{cve_id}{ID assigned by Common Vulnerabilities and Exposures (CVE)}
  \item{cve_rejected}{\code{TRUE} if vulnerability is rejected by
  the CVE editors}
  \item{published}{Date when the vulnerability got published according to
  the vulnerability entry's XML tag \code{vuln:published-datetime}}
  \item{cwe}{Common Weakness Enumeration (CWE) vulnerability type. NVD uses
  a subset of CWEs, meaning some vulnerabilities might have type \code{NA}.}
  \item{cvss_score}{Common Vulnerability Scoring System (CVSS) version 2.0
  Base Score}
  \item{cvss_av}{Access Vector (AV) [Local (L), Adjacent Network (A),
  Network (N)] used to calculate Base Score}
  \item{cvss_ac}{Access Complexity (AC) [High (H), Medium (M), Low (L)]
  used to calculate Base Score}
  \item{cvss_au}{Authentication (Au) [Multiple (M), Single (S), None (N)]
  used to calculate Base Score}
  \item{cvss_c}{Confidentiality Impact (C)
  [None (N), Partial (P), Complete (C)]used to calculate Base Score}
  \item{cvss_i}{Integrity Impact (I) [None (N), Partial (P), Complete (C)]
  used to calculate Base Score}
  \item{cvss_a}{Availability Impact (A)
  [None (N), Partial (P), Complete (C)] used to calculate Base Score}
}}
\source{
XML Version 2.0 from
\url{https://nvd.nist.gov/vuln/data-feeds#CVE_FEED} as of 7 October 2017,

NVD's CVSS introduction: \url{https://nvd.nist.gov/vuln-metrics/cvss},

A Complete Guide to the Common Vulnerability Scoring System
Version 2.0, Base-Metrics:
\url{https://www.first.org/cvss/v2/guide#2-1-Base-Metrics}.
}
\usage{
nvd
}
\description{
A data set obtained by processing NVD XML files corresponding to CVE ID
entries from CVE-2011 to CVE-2016 (included).
}
\keyword{datasets}
