% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlparser.R
\name{nvd_entry_parser}
\alias{nvd_entry_parser}
\title{NVD vulnerabilities XML parser}
\usage{
nvd_entry_parser()
}
\value{
A list containing a node-parsing-function and a constructor, where
constructor returns a \code{data.table::data.table} of entries from XML.
}
\description{
A function that goes into \code{XML::xmlEventParse} as \code{branches}
argument to parse NVD entries' nodes marked as \code{entry}.
}
\examples{
\dontrun{
  parser <- nvd_entry_parser()
  XML::xmlEventParse(file = "test.csv", handlers = list(), branches = parser)
  parser$get_vulnerabilities()
}
}
\keyword{internal}
