% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvdmodel.R
\name{NVDModel}
\alias{NVDModel}
\title{...}
\description{
...
}
\section{Public methods}{
{
\itemize{
  \item \strong{$new(cwe, cwe_ts, start_year, end_year, end_month)} creates new NVD object
  \item \strong{$getStartYear()} item
  \item \strong{$getEndYear()}
  \item \strong{$getTrainingSet()} ...
  \item \strong{$getTestSet()} ...
  \item \strong{$getFcasted()} ...
  \item \strong{$getMethod()} ...
  \item \strong{$getAssessment()} ...
  \item \strong{$getFcastPeriod()} ...
  \item \strong{$getTSetChar()} ...
  \item \strong{$getResiduals()} ...
  \item \strong{$getMergedTrainTestSet()} ...
  \item \strong{$getSignificanceLevel()} ...
  \item \strong{$areResidualsNotRandom()} ...
  \item \strong{$areResidualsNotNormal()} ...
  \item \strong{$isPiIgnored()} ...
  \item \strong{$isBootstrapNotUsed()} ...
  \item \strong{$setSignificanceLevel()} ...
  \item \strong{$setResidualsNotRandom()} ...
  \item \strong{$setBootstrapNotUsed()} ...
  \item \strong{$setPiIgnored()} ...
  \item \strong{$setFcasted()} ...
  \item \strong{$setFcastPeriod()} ...
  \item \strong{$assessModel()} ...
  \item \strong{$testResidualsRandomnessBox()} ...
  \item \strong{$testResidualsRandomnessBreusch()} ...
  \item \strong{$testResidualsNormality()} ...
  \item \strong{$getPlot()} ...
  \item \strong{$compareCombinedAccuracy()} ...
  \item \strong{$findLambda()} ...
}
}
}

\section{Note}{
 {
This class's private function `residuals_lag` is using GPL-3 licensed code from (R package "forecast"
by Rob Hyndman, Mitchell O'Hara-Wild, Christoph Bergmeir, Slava Razbash
and Earo Wang)
\href{https://github.com/robjhyndman/forecast/blob/c87f33/R/checkresiduals.R}{https://github.com/robjhyndman/forecast/blob/c87f33/R/checkresiduals.R}
from lines 122 and 123 to calculate the lag parameter.
}
}

\keyword{internal}
