% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe_analyser_utils.R
\name{separate_and_merge}
\alias{separate_and_merge}
\title{CWE organiser}
\usage{
separate_and_merge(one_type_vulns, multi_type_vulns)
}
\arguments{
\item{one_type_vulns}{Vulnerabilities with only one CWE}

\item{multi_type_vulns}{Vulnerabilities with at least 2 CWE-s: CWE-A|CWE-B...}
}
\value{
A data.table with "published", "cwe" and "cvss_score" columns
}
\description{
Organise every CWE by date.
}
\examples{
\dontrun{
vulns <- discern_types(get_type_score_between(2011, 2012))
separate_and_merge(vulns[[1]], vulns[[2]])
}
}
\keyword{internal}
