% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_maker.R
\name{vulns_into_time_series}
\alias{vulns_into_time_series}
\title{Monthly average CVSS time-series creator}
\usage{
vulns_into_time_series(chosen_vulns, vulns_by_month, start_year)
}
\arguments{
\item{chosen_vulns}{A character vector with CWEs}

\item{vulns_by_month}{Vulnerabilities with monthly average CVSS scores}

\item{start_year}{The start year of the observed period}
}
\value{
A monthly time-series object <"mts" "ts" "matrix"> of
average CVSS scores of vulnerabilities with CWEs in columns
}
\description{
Convert processed data into a time series object
}
\keyword{internal}
