% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvssforecaster.R
\name{CVSSForecaster}
\alias{CVSSForecaster}
\title{CVSS Forecaster}
\description{
Forecasts monthly mean CVSS scores of subsets of CWE vulnerabilities.
}
\section{Usage}{

\preformatted{
## Create new forecasting object.
cf <- CVSSForecaster$new()
## Find interesting CWE vulnerability categories between years 2011 and 2016.
cf$setCWENamesAndSeries()
## Set models (creates the forecasts models for the selected CWEs' time
series and measures the accuracy).
cf$setARIMA()
cf$setETS()
## Get forecast plots.
cf$getPlots(cf$getARIMA())
cf$getPlots(cf$getETS())
## Get ARIMA model's AICc, AIC and BIC.
cf$getInformationCriterions(cf$getARIMA())
## Get forecast accuracy measures.
cf$getAssessments(cf$getARIMA())
}
}

\section{Details}{

\code{$new()} creates new CVSSForecaster object
}

\examples{
\dontrun{
## Create new forecasting object.
cf <- CVSSForecaster$new()
## Find interesting CWE vulnerability categories between years 2011 and 2016.
cf$setCWENamesAndSeries()
## Set models (creates the forecasts models for the selected CWEs' time
series and measures the accuracy).
cf$setARIMA()
cf$setETS()
## Get forecast plots.
cf$getPlots(cf$getARIMA())
cf$getPlots(cf$getETS())
## Get ARIMA model's AICc, AIC and BIC.
cf$getInformationCriterions(cf$getARIMA())
## Get forecast accuracy measures.
cf$getAssessments(cf$getARIMA())
}

}
