% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvssforecaster.R
\name{CVSSForecaster}
\alias{CVSSForecaster}
\title{CVSS Forecaster}
\description{
Forecasts monthly mean CVSS scores of subsets of CWE vulnerabilities.
}
\section{Usage}{

\preformatted{
## Create new forecasting object.
cf <- CVSSForecaster$new()
}
}

\section{Details}{

\code{$new()} creates new CVSSForecaster object
}

