% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{CWE}
\alias{CWE}
\title{CWE data handler}
\description{
Manage the data that forms the basis of the forecasts
}
\section{Usage}{

\preformatted{
## Create new CWE object.
cf <- CWE$new()
}
}

\section{Details}{

\code{$new()} creates new CWE object
}

