% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvdmodel.R
\docType{data}
\name{NVDModel}
\alias{NVDModel}
\title{Lag finder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NVDModel
}
\description{
Function, using GPL-3 licensed code from (R package "forecast"
by Rob Hyndman, Mitchell O'Hara-Wild, Christoph Bergmeir, Slava Razbash
and Earo Wang)
https://github.com/robjhyndman/forecast/blob/c87f33/R/checkresiduals.R
from lines 122 and 123 to calculate the lag parameter.
}
\keyword{internal}
